+++ lib/src/internal/print_appender_default.dart
@@ -1,0 +1,4 @@
+import 'package:logging_appenders/src/logrecord_formatter.dart';
+
+LogRecordFormatter defaultLogRecordFormatter() =>
+    const DefaultLogRecordFormatter();