+++ lib/src/internal/print_appender_io.dart
@@ -1,0 +1,7 @@
+import 'dart:io';
+
+import 'package:logging_appenders/src/logrecord_formatter.dart';
+
+LogRecordFormatter defaultLogRecordFormatter() => stdout.supportsAnsiEscapes
+    ? const ColorFormatter()
+    : const DefaultLogRecordFormatter();